Trying to find out why Hotel is routing to SearchHotels and Search.
