--- BucketListAdventures/Models/Search Hotels/SearchHotelsModel.cs
@@ -1,6 +1,0 @@
-﻿namespace BucketListAdventures.Models
-{
-    public class SearchHotelsModel
-    {
-    }
-}
