+++ BucketListAdventures/ViewModels/RequiresAttribute.cs
@@ -1,5 +1,6 @@
 ﻿namespace BucketListAdventures.ViewModels
 {
+    [AttributeUsage(AttributeTargets.All)]
     internal class RequiresAttribute : Attribute
     {
         public string ErrorMessage { get; set; }
