--- BucketListAdventures/ViewModels/Search Hotels/SearchHotelsViewModel.cs
@@ -1,20 +1,0 @@
-﻿using Microsoft.AspNetCore.Mvc;
-using System.ComponentModel.DataAnnotations;
-
-namespace SearchHotels.ViewModel
-{
-    public class SearchHotelsViewModel
-    {
-        public string CityName { get; set; }
-
-        [BindProperty, DataType(DataType.Date)]
-        public DateTime? CheckIn { get; set; } = null;
-
-        [BindProperty, DataType(DataType.Date)]
-        public DateTime? CheckOut { get; set; } = null;
-        public int NumOfRooms { get; set; }
-        public int NumOfAdults { get; set; }
-        public int[] ChildAges { get; set; }
-
-    }
-}
