+++ BucketListAdventures/Views/Home/Details.cshtml
@@ -1,5 +1,4 @@
-﻿
-<img src="./images/Bucket compass.jpg" alt="Bucket compass" width="500" height="600">
+﻿<img src="./images/Bucket compass.jpg" alt="Bucket compass" width="500" height="600">
 
 @if (ViewBag.activityDetails != null)
 {
