+++ BucketListAdventures/Views/Home/Search.cshtml
@@ -1,5 +1,4 @@
-﻿
-@using SearchActivities.ViewModel;
+﻿@using SearchActivities.ViewModel;
 @model SearchViewModel
 
 <form method="post">
