+++ BucketListAdventures/Views/Hotels/DisplayHotels.cshtml
@@ -1,0 +1,10 @@
+﻿
+
+<ul>
+    @foreach (var hotel in ViewBag.hotelsObject)
+    {
+        <li>
+            <a href="@Url.Action("DisplayHotelDetails","SearchHotels")?hotel=@hotel.name">  @hotel.name </a>
+        </li>
+    }
+</ul>
