+++ BucketListAdventures/Views/Hotels/DisplayHotels.cshtml.cs
