--- BucketListAdventures/Views/SearchHotels/DisplayHotelResults.cshtml
@@ -1,10 +1,0 @@
-﻿
-
-<ul>
-    @foreach (var hotel in ViewBag.hotelsObject)
-    {
-        <li>
-            <a href="@Url.Action("Details","SearchHotels")?hotel=@hotel.name">  @hotel.name </a>
-        </li>
-    }
-</ul>
